+++ src/Dropzone/Dropzone.uni.driver.d.ts
@@ -1,0 +1,7 @@
+import { BaseUniDriver, UniDriver } from 'wix-ui-test-utils/unidriver';
+
+export interface DropzoneUniDriver extends BaseUniDriver {
+  dropFiles(files: File[]): Promise<void>;
+  getContentElement(): UniDriver;
+  getOverlayElement(): UniDriver;
+}