+++ src/TagList/index.d.ts
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { TagProps } from '../Tag';
 import { ButtonProps } from '../Button';
+import { TooltipProps } from '../Tooltip';
 import { OmitPolyfill } from '../common';
 
 export type TagListSize = 'small' | 'medium' | 'large';
@@ -10,11 +11,22 @@ export type TagListProps = {
   tags: Array<OmitPolyfill<TagProps, 'size' | 'onRemove'>>;
   size?: TagListSize;
   onTagRemove?: (id: string) => void;
+  initiallyExpanded?: boolean;
+  maxVisibleTags?: number;
   actionButton?: TagListActionButtonProps;
+  toggleMoreButton?: (
+    amountOfHiddenTags: number,
+    isExpanded: boolean,
+  ) => ToggleMoreButtonProps;
 };
 
 type TagListActionButtonProps = Pick<ButtonProps, 'onClick'> & {
   label?: string;
+};
+
+type ToggleMoreButtonProps = {
+  label: string;
+  tooltipProps?: TooltipProps;
 };
 
 export default class TagList extends React.PureComponent<TagListProps> {}