+++ src/ThemeProvider/test/ThemeProvider.tsx
@@ -11,14 +11,18 @@ function themeProviderWithAllProps() {
     <ThemeProvider
       dataHook="dataHook"
       theme={{
-        'color-00': '#FF0000',
-        'color-05': '#FF0000',
-        'color-10': '#FF0000',
-        'color-20': '#FF0000',
-        'color-30': '#FF0000',
-        'color-40': '#FF0000',
-        'color-50': '#FF0000',
-        'color-60': '#FF0000',
+        color00: '#FF0000',
+        color05: '#FF0000',
+        color10: '#FF0000',
+        color20: '#FF0000',
+        color30: '#FF0000',
+        color40: '#FF0000',
+        color50: '#FF0000',
+        color60: '#FF0000',
+        textColorPrimary: '#FF0000',
+        textColorPrimaryLight: '#FF0000',
+        textColorSecondary: '#FF0000',
+        textColorSecondaryLight: '#FF0000',
       }}
     />
   );