+++ src/TimeInput/test/TimeInput.tsx
@@ -25,6 +25,9 @@ function TimeInputWithAllProps() {
       minutesStep={11}
       width="auto"
       customSuffix={<div>hello</div>}
+      hideStatusSuffix
+      status="error"
+      statusMessage="message"
     />
   );
 }